+++ pages/experiments/[id].tsx
@@ -7,6 +7,7 @@ import ExperimentsApi from '@/api/ExperimentsApi'
 import MetricsApi from '@/api/MetricsApi'
 import SegmentsApi from '@/api/SegmentsApi'
 import ExperimentDetails from '@/components/ExperimentDetails'
+import ExperimentToolbar, { ExperimentToolbarMode } from '@/components/ExperimentToolbar'
 import Layout from '@/components/Layout'
 import { ExperimentFull, MetricBare, Segment } from '@/models'
 
@@ -17,10 +18,60 @@ export default function ExperimentPage() {
   const experimentId = toIntOrNull(router.query.id)
   debug(`ExperimentPage#render ${experimentId}`)
 
+  const [mode, setMode] = useState<ExperimentToolbarMode>('view')
   const [fetchError, setFetchError] = useState<Error | null>(null)
   const [experiment, setExperiment] = useState<ExperimentFull | null>(null)
   const [metrics, setMetrics] = useState<MetricBare[] | null>(null)
   const [segments, setSegments] = useState<Segment[] | null>(null)
+
+  /* istanbul ignore next; to be handled by an e2e test */
+  function handleCancel() {
+    // TODO: If form is dirty, then prompt for cancellation.
+    setMode('view')
+  }
+
+  /* istanbul ignore next; to be handled by an e2e test */
+  function handleConclude() {
+    setMode('conclude')
+    setTimeout(() => {
+      window.alert('TODO: Handle conclude mode.')
+    }, 1)
+  }
+
+  /* istanbul ignore next; to be handled by an e2e test */
+  function handleDisable() {
+    setMode('disable')
+    setTimeout(() => {
+      const disable = window.confirm('Are you sure you want to disable?')
+
+      if (disable) {
+        setTimeout(() => {
+          window.alert('TODO: Handle disable mode.')
+          setMode('view')
+        }, 100)
+      } else {
+        setMode('view')
+      }
+    }, 1)
+  }
+
+  /* istanbul ignore next; to be handled by an e2e test */
+  function handleEdit() {
+    setMode('edit')
+    setTimeout(() => {
+      window.alert('TODO: Handle edit mode.')
+    }, 1)
+  }
+
+  /* istanbul ignore next; to be handled by an e2e test */
+  function handleSave() {
+    if (mode === 'conclude') {
+      window.alert('TODO: save conclusions')
+    } else if (mode === 'edit') {
+      window.alert('TODO: update details')
+    }
+    setMode('view')
+  }
 
   useEffect(() => {
     if (experimentId === null) {
@@ -46,7 +97,19 @@ export default function ExperimentPage() {
   return (
     <Layout title={`Experiment: ${experiment ? experiment.name : 'Not Found'}`} error={fetchError}>
       {experiment && metrics && segments && (
-        <ExperimentDetails experiment={experiment} metrics={metrics} segments={segments} />
+        <>
+          <ExperimentToolbar
+            experiment={experiment}
+            mode={mode}
+            onCancel={handleCancel}
+            onConclude={handleConclude}
+            onDisable={handleDisable}
+            onEdit={handleEdit}
+            onSave={handleSave}
+            section='details'
+          />
+          <ExperimentDetails experiment={experiment} metrics={metrics} segments={segments} />
+        </>
       )}
     </Layout>
   )
